+++ src/twilio/.twilio-functions
@@ -1,0 +1,10 @@
+{
+	"projects": {
+		"AC2232b0eb4a28b188880935911ff2badb": {
+			"serviceSid": "ZSa1653eaa90a0bfb8842dc6dda5bca80b",
+			"latestBuild": "ZB4655a09cb4c316fe8447981e082dfd98"
+		}
+	},
+	"serviceSid": "ZSa1653eaa90a0bfb8842dc6dda5bca80b",
+	"latestBuild": "ZB4655a09cb4c316fe8447981e082dfd98"
+}
