+++ src/twilio/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "twilio-requests",
+  "name": "mutual-aid",
   "version": "0.0.0",
   "private": true,
   "scripts": {
