--- src/twilio/twilio-requests/.nvmrc
@@ -1,1 +1,0 @@
-10
