--- src/twilio/twilio-requests/.twilio-functions
@@ -1,3 +1,0 @@
-{
-	"projects": {}
-}
